Material Canvas: Fixing screen turning red from invalid entity bounds
This change fixes an issue with the material canvas and material editor viewport turning red momentarily on startup and anytime the viewport model is invalid or deleted. This is caused because requesting the object entity local and world bounds returns an invalid bounding box if no model is loaded.

This change also renames a handful of variables and functions.

https://github.com/o3de/o3de/issues/10879
Signed-off-by: gadams3 <guthadam@amazon.com>