Tweak DPE patch path normalization behavior
I received some API feedback about patch handling within the Document Property Editor, namely that the prevalent use of `EndOfArray` entries in our `GenerateHierarchicalDeltaPatch` algorithm (and the validity of its prevalence) led to significant friction for consumers handling patches downstream, as to fully resolve an `EndOfArray` path requires a denormalization operation via a lookup in a provided `Dom::Value`, which is both cumbersome and potentially expensive. While there still remains a valid and open architectural question as to whether JSON patch is the right abstraction for the DPE, this patch attempts to address some of this friction in the short term.

To that end:
1. Our `EndOfArray` usage has been audited and restricted to only the one place where it's called out (and indeed mandated) in the JSON patch spec: `Add` operations, and `Add` operations only, denoting an insertion to the end of the target array.
2. An additional flag for patch generation has been added, `m_generateDenormalizedPaths`, in which we provide nonconformant paths for `Add` operations specifying an index past the end of the target array or node. This is a breaking change for serialized patches, but a non-breaking change for our internal patching algorithms, which have been softened to allow for this case.
3. Additionally, in the interest of allowing DPE consumers to continue to provide JSON patches in whatever form is most convenient, patches provided to an adapter are explicitly denormalized before being sent to any consumers (such as proxy adapters or the DPE widget itself).

The intent here is to make real, concrete, addressable paths *always* available to downstream systems to prevent unneeded and cumbersome book-keeping.

Signed-off-by: Nicholas Van Sickle <mail@nickvansickle.com>