Improve Asset Manager Reload behavior. * Changes AM::ReloadAsset() to return the newly-created asset, so that callers can use it to monitor the state of version getting reloaded. * Changes AssetCommon::Reload() to immediately change the current asset to the one getting reloaded. * Changes AssetInternal::GetAssetData() to get any queued reloaded data as well as currently-loaded data, though at a higher priority.
The net effect of these changes is that calling myAsset->Reload() will now change the myAsset reference into a queued state and clear its current set of data, as opposed to leaving it in a Ready state with valid data. The previous logic meant that myAsset wouldn't know anything about the reloading, and the caller would need to listen for OnAssetReloaded to *then* change myAsset to the new reloaded asset reference.

The more subtle change is the one to GetAssetData(). This changes how BlockUntilLoadComplete() works with reloaded assets. Previously, if you called myAsset->Reload(); myAsset->BlockUntilLoadComplete(); the second call wouldn't actually block because it would see that myAsset is currently loaded (since it is) and would return immediately. With this change, the call will get the reload asset data instead and will block until the reload is complete.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>