[SceneCore] Work around memory leak in GetNodeContext behavior reflection
This method used to allocate and return a pointer to a `GraphObjectProxy`
object. The trouble is, BehaviorContext has no way to express that this
object should be owned by the script, and deallocated when the script's
variable goes out of scope.

This works around that problem by returning this type by value. The script
always owns value types returned from methods.

To allow for this, the `GraphObjectProxy` must have a default constructor.
This is because the script context implements by-value return types by
allocating space for the return type before the call is made, default
constructing the resulting memory, then using copy assignment from the
return value after the call is made. If the default constructor is not
present, the copy assignment method may fail, since it is being run on an
uninitialized object.

Fixes #11827

Signed-off-by: Chris Burel <burelc@amazon.com>