[ManipulatorTestFramework] Fix memory leak of manipulators in tests
These manipulators are stored in shared_ptrs. The manipulator itself can
register a callback function, and the tests were creating a lambda with
a copy of the shared_ptr as the callback. Thus the object stored a function
that maintained ownership of the object that owned the function. This
circular reference resulted in the manipulator never beeing freed.

Signed-off-by: Chris Burel <burelc@amazon.com>