Fixed the Color Picker color space for materials and lights.
Originally the color picker was set to be in sRGB, and configure a Linear sRGB property. This PR https://github.com/o3de/o3de/pull/6812 change that so the color picker and the property were both treated as sRGB, to make the numbers match. It's understandable that the developer thought this was an improvement, but this change was not actually the correct thing to do. The most common and natural workflow is for artists to work in a sRGB color space, while the material and lighting systems store color values as Linear sRGB. That is why the color picker needs to be configured as it was originally.

In the long term, we would like to expand the color picker to allow users to select what color space they want to operate in. It may be sRGB by default, but the user could put the color picker in any mode they want. And the output property value will always be linear. In the short term, we plan to make some minor additions that would hopefully alleviate some of the confusion cause by the differing values.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>