This is a first pass at making the Color Picker experience more clear.
Since the Color Picker operates in sRGB and the property being edited is Linear sRGB, the numbers do not match between Color Picker and the Property Inspector. This is intentional, but admittedly confusing as seen in these issues:
- https://github.com/o3de/o3de/issues/4210
- https://github.com/o3de/o3de/issues/2980
- https://github.com/o3de/o3de/issues/3736

So I'm adding some alternate read-only fields to the Color Picker that will display the final output value in the destination color space. These values will match the value shown in the Property Inspector.

The implementation in ColorPicker isn't ideal, because the client code is responsible for doing the conversion and populating these values every time the main color changes, using the currentColorChanged callback. It would be better for ColorPicker to manage the color conversion internally. But that level of change is out of scope for the upcoming release. There has been talk of wanting to make a more significant investment in ColorPicker, to enable it to operate in any color space of the artists choosing. Making those changes will naturally address this concern as it too will require making ColorPicker do the conversions internally.

Also...
- I moved the controls around a bit to group the float edit boxes closer to the sliders, which I think is more intuitive. Similarly, the colorspace comment text is located together with the alternate colorspace read-only fields.
- The comment text used to also serve as a label for the float edit boxes, now it is purely a colorspace comment, which says "Color space: sRGB" by default.
- Fixed a potential issue where setConfiguration could be called and not actually change the m_configuration.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>