Updated to Gem CMake targets templates to reference the "gem_name" from the gem.json
This changes makes it easier to rename a gem via updating it's
"gem_name" field in the gem.json file.

Previously users would also need to update the CMakeLists.txt to update
the target names to match the gem name in order for CMake to find the
`<gem-name>.<variant>` target that it checks before adding the gem module
filename to the generated `cmake_dependencies.<target>.setreg` which
indicates that the gem is active.

closes #12309

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>