Material Canvas: Update sorting logic to force positioning of input and output nodes
Inspector and code generation now use the same sorting logic that forces input nodes to the beginning and output nodes to the end of the container. This resolves problems where disconnected nodes and slots with multiple connections can result in nondeterministic ordering.

Also implemented custom trace print functions that only printed the condition is true to help reduce repeated checks and blocks. We’ll discuss if these are worth moving with the rest of the core trace functions in macros.

Signed-off-by: gadams3 <guthadam@amazon.com>