Add support to the JSONImporter to replace any aliases in the import path before appending it to the import absolute path
This allows the `ResolveImports` function to be used with a path that
starts with a path that has the form of
`@gemroot<gem-name>@/rest/of/path@` and smartly update the import path
based on whether it the import filepath resolves to an absolute path in
which case it replaces the path importAbsPath value. Or if the import
path doesn't contain a fileio alias and is relative, it is appended to
the importAbsPath value.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>