Fix crash on editor startup due to missing assets
When starting editor from a cache that does not contain an assetcatalog.xml, editor relies on network messages from the AP to build its in-memory catalog.  If the asset processor/editor are restarted before the first time starutp has completed, AP will resume where it was in the processing without resending the previously completed assets to the editor.

This fixes AP to send out all the existing assets in the catalog when a new connection is established.
Also fixed issue in a few usages of CompileAssetSync where the Unknown return case was not being handled, which can occur when an asset has already been compiled.

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>