Adding ISO 8601 timestamp utility functions
The ISO 8601 timestamp utility functions support the extended format
which includes characters such as as dashes(`-`) and colons (`:`).
So the timestamp would like the following "2022-10-10T12:00:00Z"

The ISO8601 basic format support has been added as well.
A timestamp in that format would look as follows "20221010T120000Z"

Finally a ISO 8601 like format has been added to support a mix of the
extended format for human readability in the Date portion and the basic
format for filesystem compatibility with Windows in the Time portion
in order to support timestamps that can be used as part of filenames.
The Windows Kernel doesn't support paths with colons in them.

That format current looks like the following: "2022-10-10T120000Z"
That format is up for debate and feedback to change it is welcome.

Only support for the Date and Time portion of the ISO8601 format is
being added.
The ability to get
[Duration](https://en.wikipedia.org/wiki/ISO_8601#Durations) or [Week
based](https://en.wikipedia.org/wiki/ISO_8601#Week_dates) timestamp support
is not added at this time.

Switching to C++20 would alliviate the need for this functions since the
chrono date library support stream operators for outputing ISO 8601
timestamps.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>