Fix insertion into NodePaletteAutoCompleteModel
In NodePaletteAutoCompleteModel::AddAvailableItem there was an incorrect
call to QAIM::beginInsertRows.
Here, we wan to append the new item as the last row. According to Qt
documentation [1], in such case we should pass the index one past the
current last index. Thus, the "-1" were incorrect.

For example, when inserting into an empty model we would call
`beginInsertRows(QModelIndex(), -1, -1)` which triggers an assertion
in QAIM::beginInsertRows().

[1] https://doc.qt.io/qt-5/qabstractitemmodel.html#beginInsertRows

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>