Significantly reduces memory usage of Editor (#11860)
This change makes RapidJSON no longer pool its allocations.  Instead it
directly asks the AzCore SystemAllocator for memory.  The AzCore
SystemAllocator already pools memory in 64k blocks itself.

The memory savings are roughly 2x for any time JSON is involved.
They come from the fact that RapidJSON would ask for
64K bytes PLUS 32 extra bytes for a chunk header.  This would cause
the AzCore SystemAllocator to have to ask for 2x 64k chunks since the
data would not fit in 1x 64 chunk.   What makes it worse is that
this would get doubled again by the fact that AzCore asks for it
to be 64k aligned, which doubles the amount because of the way
__malloc_aligned functions.

Allocating X bytes with Y alignment on Windows will cause
__malloc_aligned to allocate at least X + Y actual bytes, since that
is the only way to ensure that there is a contiguous block of memory
within the returned pointer that starts at a Y alignment but still has
X bytes after it.

The savings here come from the fact that now all the small objects
like strings and tokens and nodes and so on will all go into the AzCore
chunk allocator (HPHASchema) in the buckets meant for small objects,
which are combined together in a pooling mechanism, without the
'double chunking' occuring.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>