The new AzCore/Debug/PerformanceCollector (#12551)
What does this PR do?

Added a new class to AzCore/Debug called PerformanceCollector. It is a general purpose class that helps measure and record performance data. The output data is generated using AzCore/Metrics/JsonTraceEventLogger which produces JSON files in the Google Trace data format.

For details on how to use take a look at the file AzCore/Tests/Debug/PerformanceCollectorTests, but here is the gist of it:
PerformanceCollector provides two mechanisms to measure code performance in microseconds:

    Scope based:
    Use the companion AZ::Debug::ScopeDuration class. Let's say you want to measure the duration of a particular function:

void MyClass::MyFunction()
{
     ...  Some Code ...
}

Then you can do (Assuming MyClass owns or has access to a PerformanceCollector object.

void MyClass::MyFunction()
{
     AZ::Debug::ScopeDuration scope(m_performanceCollectorPtr, "MyParameter");
     ...  Some Code ...
}

In the code above, the duration of MyClass::MyFunction() will be recorded under the parameter named MyParameter.
Equivalently, You can use AZ::Debug::ScopeDuration to measure any other block of code.

    Periodic events/functions:
    This option is useful to measure the time it takes between calls to periodic functions like OnSystemTick() or OnTick():

void RPISystemComponent::OnSystemTick()
{
    m_performanceCollector->RecordPeriodicEvent("MyCpuTime");
    ... code ...
}

In the code above, the time it takes in between calls to RPISystemComponent::OnSystemTick() will be recorded in the parameter named MyCpuTime.
About size of the output data.

The size of the output data will be drastically different depending on the output mode you choose when configuring the PerformanceCollector.

    DataLogType::LogStatistics: This is the default mode of operation. Each measured parameter will produce a single json line for all the data collected under an arbitrary number of frames. Whether the data is collected across 1 frame or 10000000 frames, only one line of json data with a statistical summary will be produced. Example:
    {"name":"Graphics Render Time","cat":"Graphics","ph":"X","ts":1665585433079433,"pid":7228,"tid":51260,"args":{"avg":14802.554166666678,"min":13290.0,"max":16197.0,"sampleCount":1200,"units":"us","variance":418435.9353454264,"stdev":646.866242236698,"mostRecentSampleValue":14840.0},"dur":14802},
    DataLogType::LogAllSamples: This is the verbose alternative. For each frame/sample of recorded data a json object will be added to the output file. This means that if you choose to capture performance data for a single parameter across 10000000 frames, then 10000000 json objects will added to the output file. Use this mode judiciously.

PerformanceCollector & RPISystemComponent.

In this PR the first customer of PerformanceCollector class is the RPISystemComponent. The RPISystemComponent now exposes the following CVARs:

    r_metricsDataLogType=statistical [ ]
    r_metricsFrameCountPerCaptureBatch=1200 [ ]
    r_metricsNumberOfCaptureBatches=0 [ ]
    r_metricsWaitTimePerCaptureBatch=0[ ]
    When the user changes r_metricsNumberOfCaptureBatches to be greater than 0, it will trigger performance collection for the following parameters: Graphics Simulation Time, Graphics Render Time and Engine CPU Time.

Why? How does this compare to AzCore/Debug/Profiler and AZ_PROFILE_* macros?

AZ_PROFILE_* Macros are the de-facto and encouraged method to place profiling markers across the code base. It produces data files that can be used with PIX to debug and root cause performance issues. See this wiki for more details: Profiling Tools.
PerformanceCollector, on the other hand, is useful to keep track of key performance metrics, that can be collected regularly and help us spot regressions or improvements. It is easy to use at runtime without requiring special compilation flags. It leverages the Google Trace JSON format because it uses the JsonTraceEventLogger APIs.
How was this PR tested?

Added a Unit Test at AzCore/Tests/Debug/PerformanceCollectorTests.cpp which passes on Both Linux & Windows.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>