Improves AssetProcessorBatch startup time
Also fixes a unit test that was always failing on mac (nobody was
running the unit tests on mac) due to the way temp folders may be
inside a symlink.

I also replaced most scan folders in the config file with a glob
version instead of a regex version, which is significantly faster.

I also eliminated legacy excludes that are either covered by another
exclude (for example the temp hold folder is already covered by the
broader temp exclude) as well as those that just don't exist anymore
such as the animation compression folder (from legacy CryEngine).

This will also improve the Asset Processor GUI version startup time.

This set of changes come from profiling the AP startup time when the
cache is already mostly populated (Zero Analyis Mode).  It optimizes the
functions that were consuming most of the time and uses the file list
gotten from the scanner to skip having to compute the excluded folders
or excluded files more than once.

This was tested by running the AP Batch version with --zeroAnalysisMode
parameter, as well as --project-path parameter to point it at
AutomatedTesting, as well as running the automated tests.

I also ensured that the number of files returned from the scanner
is the same before and after this change, to ensure that the globs
were equivalent to the regees.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>