Fix test logic in EMotionFX
The call to create context menu (`OnContextMenuEvent`) was called at the
same location. However, during subsequent calls when building the menu
there was already a node at the click point from the first iteration.

Menu built this way didn't actually contain the action to create a node.
However, since the old menu was still there (because this menu is not
created through a blocking call) the action was still found.

This wasn't a problem until the palette in the menu was migrated to
`GraphCanvas::NodePaletteWidget`. This widget should not be re-used and
it handled multiple selection calls incorrectly when in context menu.

Now, each "click" is performed at a different place and we also
do one event loop iteration in our loop to make sure that the menu from
the previous iteration is gone.

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>