Removed special case for unit test on linux.
This test was relying on the GetJobsBySourceName query to return whatever file had a matching relative path in the database along with the fact that on linux, this test had the same casing for both the overrider and original file, resulting in the same fingerprint and no work being done.
Since we're no longer expecting relative paths to be unique, the query now uses rel path + scanfolder Id, meaning the GetJobsBySourceName query returns no results when the override file is deleted in this test.

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>