Removed duplicate test that was failing on Linux.
The removed test was looking for the same information that the tests in this file look for:
https://github.com/o3de/o3de/blob/development/AutomatedTesting/Gem/PythonTests/assetpipeline/fbx_tests/pythonassetbuildertests.py#L127
But the removed test required the Editor and AP GUI to launch due to its setup, even though it intended to only use the AP Batch.

The test in pythonassetbuildertests.py runs on Linux correctly. The removed test fails on Linux.

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>