Added performance metric:Render Pipeline Gpu Time.
Added class GpuPassProfiler, that is based on the existing
ImGuiGpuProfiler but did not make changes to ImGuiGpuProfiler
at this time to simplify the size of the PR.

The core function is in
GpuPassProfiler::MeasureRenderPipelineGpuTimeInNanoseconds().
Use judiciously, as MeasureRenderPipelineGpuTimeInNanoseconds() can itself
affect performance when called. For example, if a scene is running at 300fps, calling
 MeasureRenderPipelineGpuTimeInNanoseconds(), each frame, can reduce the FPS to ~265fps.

Added CVAR r_metricsEnableRenderPipelineGpuTime, which defaults to
false, as it is the only parameter that can be expensive to measure.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>