Atom Tools: Deferring open document actions to the system tick bus and fixing other timer usage
Problems were occurring when opening documents from shortcut and menu actions. Modal message boxes can be displayed for documents that log warning and error messages. Opening a document triggers notifications that could subsequently cause the actions that triggered them to be deleted as part of menus refreshing. By deferring the open document commands to use a timer or tick bus, they will be invoked as part of the regular system tick instead of nested inside of the action callback with the potential warning or error message boxes.

Replaced all QTimers that could not be connected to a parent object for lifetime management with system tick bus.

The callbacks bound to most timers and the tick bus were updated to not capture the this pointer.

Some direct function calls were converted to bus calls so the this pointer did not have to be captured.

Created and updated test scripts that do exhaustive document opening and closing to make sure things are being cleaned up correctly.

Signed-off-by: gadams3 <guthadam@amazon.com>