EMotionFX: Fix crash in Parameter Window
When doing a full clear, the UpdateInterface() could be called with
m_selectedParameterNames not updated because this field is only
updated when selection isn't locked. However, by the end of Reinit(),
this field could keep a stale name of a parameter that has been
actually deselected. When doing a clear of all parameters, this
caused the UpdateInterface to leave the edit UI in the inspector
for a parameter that no longer exists. Clicking "Apply" in such UI
caused a crash.

Now, we update the selection arrays at the end of Reinit so
UpdateInterface has proper value of m_selectedParameterNames.

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>