Additional clang for Window fixes when using clang 15.0.2
The variadic `__VA_OPT__` preprocessor magic is available when using the Windows
clang prepreprocessor, but doesn't correctly handle the case of back to
back empty macro parameters. such as `AZ_VA_NUM_ARGS(,)` which should be
2, since there are 2 empty arguments.

The next fix is when compiling a reverse_iterator with a template
parameter of the AZStd red-black tree iterator type.
The MSVC STL implementation is trying to use `noexcept` to disable the
function in C++17, but clang treats errors in noexcept as hard errors.

Finally added a CMake Preset that can configure the engine to build for
Windows using the ninja build system and targeting the clang compiler.
The preset requires that `ninja`and `clang` are both in the PATH.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>