Added support to the JsonTraceEventLogger for querying settings
Introduced a SetName/GetName function to the IEventLogger interface that
allows associating a name with any derived event logger class

The name of the EventLogger is used as a part of a settings registry key
`/O3DE/Metrics/<Name>` to contain settings associated with the running
state of that event logger instance

The JsonTraceEventLogger will query the settings registry for the Active setting associated with the event logger name.
The query will occur from `/O3DE/Metrics/<Name>/Active`

If the event logger name changes than the Active setting is queried
using the name from `/O3DE/Metrics/<NewName>/Active`

If the setting does not exist in the settings registry, the event logger
is set to always be active.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>