Adding a ParseConfigFile function to AzCore
The ParseConfigFile function can parse a Windows style INI file without
any dynamic memory allocations and is suitable to be used in early
application startup before O3DE allocators are available.

The SettingsRegistryMergeUtils MergeSettingsToRegistry_ConfigFile
function has been updated to use the ParseConfigFile function.

Added a fix to the Comment Prefix Filter functions.
If a line contained multiple comment start characters such as `";#"` and a commented line of `# This is a comment; More text`,
then the Comment Prefix Filter function would only remove up to the
first comment start character that was found (in this case it remove up
to the semicolon and not the number sign)

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>