Allow users to configure the depth of the captured stack for allocation tracking
Now that the OSAllocator is tracking allocations, the overhead of capturing
the callstack has grown so that tests no longer finish under the specified
timeout of 15 minutes on the Linux CI machines. To mitigate this, the user
can now configure how many stack levels to capture, as a compile-time
constant. It is exposed via a CMake cache variable to enable developers
easy access to its value. The default value is 3, which should be
sufficient to identify immediate locations where allocations occur (looking
past any allocator wrappers like ChildAllocatorSchema), without too high of
an impact on performance. Of course, because tracing is disabled in release
builds, it already has no impact on those.

Signed-off-by: Chris Burel <burelc@amazon.com>