Implement BreadCrumbs in AssetBrowser
These breadcrumbs show the relative path of a selected folder or, if
an assets is selected, its folder.

Breadcrumbs are for now hidden behind the ed_useWIPAssetBrowserDesign
cvar.

When the width of Asset Browser window falls below a certain threshold,
it goes into "narrow" mode, breadcrumbs are then shown below the
search bar, view switchers and the tools menu.

Fixes: #13236

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>