Added performance metric:Frame Gpu Time. (#13070)
* Added performance metric: "Frame Gpu Time".

Added class GpuPassProfiler, that is based on the existing ImGuiGpuProfiler but did not make changes to ImGuiGpuProfiler at this time to simplify the size of the PR.

The core function is in
GpuPassProfiler::MeasureGpuTimeInNanoseconds(). Use judiciously, as MeasureGpuTimeInNanoseconds() can itself affect performance when called. For example, if a scene is running at 300fps, calling
MeasureGpuTimeInNanoseconds(), each frame, can reduce the FPS to ~290fps.

Added CVAR r_metricsMeasureGpuTime, which defaults to false, as it is the only parameter that can be expensive to measure.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>