Material Canvas: Proposed changes to remove material hot reload notification bus
Material editor, material canvas, material instance editor, and thumbnails all use the material component to load, reload, and display material previews in their viewports. These systems exercise material hot reloading more than any others besides ASV samples and tests.

Changes were previously made to reenable material finalization by default in the material builder. That allowed systems that manage materials to be able to rely on standard asset ready and reload notifications. The material component uses the asset bus for more conventional loading and hot reloading at the high level. It requires management and access to the most recent version of material assets to enumerate, get, and set materials and properties.

Other changes were also made for the instance database to be able to create and differentiate instances for different versions of assets with the same asset ID.

Those changes unblocked several issues but some still remain.

With the material reload notification bus logic still in place, competing systems are attempting to load materials and shaders.

Anytime material inputs are changed in a material graph, the material type and its assets need to be reprocessed, which sometimes results in asserts and errors related to shader resource group and data comparisons.

I suspect that those issues are caused by the instance database. The material, material type, and shaders may reload correctly and send the correct asset reload notifications but the instance database could be holding on to old versions of shaders and block new versions from being created when requested.

Removing the material reload notification bus and the self-reloading of materials, material types, and material assets has stopped most of the problems, while still preserving material hot reloading in the material component.

Another change was recently made to the material assignment class. This change sends a notification every time material instances are rebuilt. This code was crashing because it dereferenced a null instance pointer. The material instance pointer was null because the instance database function failed to create an instance, likely due to a collision with a pre-existing random ID.

The terrain feature processor would need to switch to conventional asset notification handling to hot reload materials.

Signed-off-by: gadams3 <guthadam@amazon.com>