Defer creation of the Core Metrics Logger until first game tick
Added support for controlling the cration of the core_metrics.\*.json
file via a setting key.
The Boolean setting of "/O3DE/Metrics/Core/CreateLogger" controls the
creation of the Core Metrics logger.
If the settings is missing the Core Metrics Logger will be
created by default in non-release builds and not created in release
builds..

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>