Fixes for failing unit tests more time saved on startup
Saves a bit more time by not initializing the asset view models
unless the tab is opened the first time.  Often nobody ever opens
the tab and it was pretty expensive to populate.

In addition, fixed some unit tests that failed becuase
FindFirstMatchingFile was not tolerant of absolute paths being fed
to it.  re-profiled and the time cost of those new checks (isAbsolute)
is negligible as it does not actually check the file system, is pure
string operations.

On this PC this brings the total 0-to-idle time for AutomatedTesting
as follows
```
4.7s - loading dlls, the exe itself, initializing qt widgets,
       waiting for 1 instance of AssetBuilder.exe to start.
1.7s - scanning
5s - analyzing
1.1s - waiting for second instance of assetBuilder to start
1.6s - createjobs

total: ~14s to idle  down from the original ~40s-ish.

This could be further reduced if the number of files that had
duplicates in several folders (seedlist.json, etc) were elimiated
as they cause always-createjobs.

```

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>