Fix stack-use-after-scope error in SerializeContext::EnumerateInstance
EnumerateInstance uses the reflected class element's RTTI helper to resolve
a class's UUID, using the `GetActualUuid()` method. `GetActualUuid()`
returns the `UUID` by _value_, which is stack allocated, but the
EnumerateInstance method took a reference to that stack allocated value.
It then used that now destroyed stack allocated value, and even stored a
pointer to that stack allocated value in the `DbgStackEntry` struct.

Signed-off-by: Chris Burel <burelc@amazon.com>