Fix heap-use-after-free in NameDictionary destructor
The NameDictionary stores a map of
Hash(aka uint32_t) -> ScopedNameDataWrapper. It is the
ScopedNameDataWrapper object that owns the actual NameData. During the
ScopedNameDataWrapper's destructor, it expects its internal NameData
pointer to still be valid. But the body of the NameDictionary's destructor
was destroying the NameData, without removing the ScopedNameDataWrapper
from the dictionary. The resulting call sequence was:

```
~NameDictionary()
  +--~NameData()
  \--~ScopedNameDataWrapper() (dereference deleted NameData*)
```

This avoids the use-after-free error by removing the ScopedNameDataWrapper
from the dictionary's map before deleting the NameData.

Signed-off-by: Chris Burel <burelc@amazon.com>