Fix hash_table::insert for iterators that return a temporary
`operator*()` for any iterator type is not guaranteed to return a `T&`. A
lot of times it does, but with range views, it will be increasingly likely
that it does not. Specifically, for like:

```
AZStd::views::transform([](const char elem) -> char { return elem + 1; })
```

dereferencing the iterator will invoke the lambda, and return a newly
constructed `char`, not a `char&`.

The reason this matters is because of the rules governing the lifetime
extension of a temporary. If this code was merely:

```
const key_type& valueKey = *first;
```

then it wouldn't be an issue. The return value of the lambda would be bound
to a const lvalue ref, so its lifetime gets extended. But by passing the
return value of the lambda to a function, even if the function is:

```
const key_type& key_from_value(const value_type& value) { return value; }
```

now that function call is in the way:

> a temporary bound to a reference parameter in a function call exists
> until the end of the full expression containing that function call: if
> the function returns a reference, which outlives the full expression, it
> becomes a dangling reference.

In order for this code to avoid using the `valueKey` as a dangling
reference, it must store it by value.

Signed-off-by: Chris Burel <burelc@amazon.com>