Fix path width calculation in BreadCrumbs
BreadCrumbs::fillPath used a crude approximation for how much space will
the path string take based on font metrics of a plain text string of the
path.

The actual html that was put into the QLabel was very different. It used
different spacing, could add icons, etc.

In some situations it could cause a crash because the html inserted into
the label could be longer that the available space. This caused a resize
that again, caused the re-calculation of the path. This time there was
enough space so a path portion was added again. This led to infinite
recursion and stack overflow.

Instead, the path label has the horizontal policy set to "Ignore". We
ourselves make sure to use the available space in the best way.

The calculation of the path width was also improved. Now it properly
takes the separators ("  >  ") and icons into account. Thus, the
OptimalPathWidth config option became redundant.

QFontMetrics::width was replaced with the non-deprecated
QFontMetrics::horizontalAdvance.

The showing/hiding of the menu button is now handled by ::resizeEvent().
We first calculate if the whole path will fit, and if not, we show it.

Some additional improvements:
 - configuration variables are now static so they have internal linkage,
   no need for them to be external
 - Rich text setting is explicitly set on the label, without it, there
   could be a situation where autodetection didn't work
 - any operation that can influence the total width of the path now
   calls updateGeometry() to make sure that layout is re-calculated

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>