Change `ActorComponentRequests` mutex type to `recursive_mutex` for Linux support
The `Attach_Detach_Entity` test uses the `AttachToEntity` method of the
`ActorComponentRequestBus` to make an attachment. The `ActorComponent`
itself connects to the `ActorComponentNotificationBus` of the target
entity, which tries to acquire the same mutex that's already been acquired
by the `AttachToEntity` call.

This just replicates the behavior of Windows by using a `recursive_mutex`
for that bus.

Signed-off-by: Chris Burel <burelc@amazon.com>