Remove bogus test data in the range remapper tests
This tests what happens when the range remapper is set to remap from
(-FLT_MAX,FLT_MAX) to (-FLT_MAX,FLT_MAX). In theory it should be a no-op,
but on Linux it ends up evaluating to `nan`. It is such an uncommon
scenario that it's not worth making per-platform support for this case.

Signed-off-by: Chris Burel <burelc@amazon.com>