Material Canvas: Renaming material template nodes to material output nodes and support for material graph node templates
Renaming material template nodes to material output nodes to solidify that they are the output of the graph. This also reduces some confusion with the overloaded term template.

Added support for material graph no templates, which are just predefined node configurations to use as a starting point for creating new nodes.

Signed-off-by: gadams3 <guthadam@amazon.com>