Fix style application in EMotionFX
For some reason the Editor.qss stylesheet is not applied to the
NodePaletteWidget when it's inserted into the dock widget in EMotionFX.

This widget is later reparented to the chain originating in the editor
main window so in theory it should be applied and the palette for the
tree view delegate should be proper.

As a workaround, we apply the style directly.

Fixes: #13235
Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>