Allow exclusions in filewatcher and fix filewatcher UnitTests
The filewatcher now has the ability to exclude folders and files.
This allows it to ignore the cache folder for example, and use
less resources on linux (which uses up a watch handle for every
folder it watches in the recursive subtree!).

The unit tests were also flaky and not working for linux and mac,
so as part of this, I also fixed them up.

I ran 1000 iterations of the tests (including periodic) on shuffle
and repeat on each platform before continuing.  The ones which
stress the file system are marked as periodic since they can burn SSD
due to rapid file creation/destruction and take a second or two
compared to the rest which all finish in milliseconds.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>