Remove redundant azmalloc, azcreate, and azcalloc overloads
The allocation records no longer fill in allocation names that are supplied
by the user. In many cases these names would come from AZStd containers,
and were not that useful. It is more informative to track the location that
the allocation originated from, which is captured in the record's
callstack.

Signed-off-by: Chris Burel <burelc@amazon.com>