Remove disabled AllocatorManager tests
This test was largely removed in 947adc0248d9be1bac2565d1a67d081b0d1677c7,
when the override shims were removed. The only thing that's left is in
SetUpAllocatorManagerTest, which is never called. The allocators now
register themselves with the manager on first use, and can't be removed,
so this test is no longer applicable.

Signed-off-by: Chris Burel <burelc@amazon.com>