Improve selection logic in AssetBrowserTreeView
If we're selecting a new entry or product, don't first call ::clear() on
the selection model and then select but rather use
QItemSelectionModel::ClearAndSelect.

When ::clear() is called we get a spurious current index change (to an
empty one) that is immediately followed by a new selection. This is
problematic for dealing with breadcrumbs in asset browser because we
would push an empty path.

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>