Fix breadcrumbs re-setting the current index
When file was selected the folder in which it was located was pushed to
the breadcrumbs which fired pathChanged signal and asset browser reacted
to it by selecting the folder.

Now, when we get a pathChanged signal from breadcrumbs that tells us to
select a folder that our current file is located directly, we ignore it.

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>