Make Asset Browser Breadcrumbs editable
Some additional margin is added to the BreadCrumbs in the Asset Browser
because now, when there's a frame it appeared "stuck" to the side of the
window. It's needed only in the "narrow" view because when it's together
with the search field and buttons in the wide mode, it's the layout
that provides spacing.

Fixes: #12854

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>