Implement editable breadcrumbs
This is needed in situations where the user may want to copy the current
path (e.g. to share with a colleague) or paste a saved one.

BreadCrumbs now derives from QFrame because plain QWidget doesn't allow
styling the border. And we need a border to indicate editable
breadcrumbs.

A small margin was added to the internal BreadCrumbs layout to avoid the
text being stuck to the left side of the border if breadcrumbs are
being edited.

Drive-by: sort Qt includes in BreadCrumbs.cpp

Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>