Implemented changes to make it easier to rename projects
The DefaultProject and Minimal project template now queries the
"project_name" from the project.json file to use a prefix when enabling
gems.

And `o3de_find_ancestor_project_root` function has been added to locate
the nearest project.json from a specified input directory or above

The SettingsRegistry.cmake has been updated to no longer attempt to
query a target file directory from prefix argument used as a part of the
`cmake_dependencies.<prefix>.<target>.setreg` generation.
The "prefix" argument is not necessarily a target and is often the name
of the project from a project.json file.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>