Switches the Batch AP to use the file cache too
This was achieved by making the AP notify when an intermediate file
was changed.  note that the notification happens instantly, so it would
be even more reliable than the file watcher notifying.

The file watcher will also notify though.  This doesn't cause issues.

Removed a test which asserted that creating a file anywhere in Assets
should immediately have it visible in the file cache.  This is only
ever true if using a pass thru file cache, which we no longer use.

In all other real world cases, this is only true for Intermediate, as
even the fast OS file notifier is async and takes a moment to notify.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>