Added `GetUnsafeAttributeReader` function to AZ Attribute
The `GetUnsafeAttributeReader` virtual function returns a wrapper
`UnsafeAttributeInvoker` struct which wraps an AttributeReader in a
temporary AttributeInvocable that can forward to another
AttributeInvocable except that it accepts its first parameter as a `void*` instead of the class type.

This allows using lambdas wrapped in AZ Attributes that accept a
non-void pointer argument to be used by code with type erased class
instances(such as Property Editor Widgets).

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>