Fixes a crash on shutdown and removes a useless test
Crash on shut down fix details:  The asset processor would
listen to its own update messages once it finished processing its
assets. This caused it to try to update its source/product model
for its tree view.  But with these batch of optimizations it does
not populate the tree view until you actually go looking at it, so
it would not exist at that moment.

Remove useless test details:  A test created a dummy file on disk
and then immediately asserted that the file cache knew about the file.

This is not a useful test, as the real file cache doesn't actually
know about any given file until something tells it about the file,
for example, the scanner or a direct notify.   The only way this test
was passing before is that the file cache was a pass-thru cache in test
mode, which just does a direct file op without caching anything (kind
of like a dummy/mock file cache).  So this test wasn't testing much
except that operating system file exists call worked.  Now that the
file cache is "real" and no longer a mock or dummy that just passes
operations thru to the OS, we'd have to inform it that the file exists
via watcher or direct call, which also wouldn't test much.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>