Small startup performance improvement for AssetProcessor
This function is one of the very hot-path functions when AP is doing
its fast scan.

QStrings cache the length of the string - so we don't need to compare
the watch folder unless they're exactly he same length.  This saves
about 5 seconds (22 down to 17) in a full run on multiplayer sample.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>