Removed the broken seed builder. (#13975)
This was used originally to facilitate asset relocation: It checked if any seed lists referenced the asset to be relocated.
However, it didn't work in practice. Often times asset seed files use the same relative file name for the default seed system, which does not play well with the asset system. If multiple source assets exist at the same relative path to the scan folder root for the scan folder they are in, the Asset Processor selects only one of them to be allowed to exist, which means that most seed lists (the default seeds) ended up invisible to this system.
Testing it farther locally, it wasn't actually generating dependencies correctly, the jobs weren't running correctly, and the asset relocation wasn't catching these references in the seed list file to the file I attempted to relocate.

To address this, we'll need to re-evalute and likely need to follow a different approach for resolving references from seeds to relocated assets, with the current relocation system.

Also note that the new relocation system (see RFC https://github.com/o3de/sig-content/issues/77) will likely require us to revisit this entire asset relocation system in the asset processor batch. We may discard it, if we find the new asset relocation system doesn't need any additional tooling.

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>