Updated the old Archive System handling of the OnSystemActivated lifecycle event
Previously it would clear out the list of Archives with asset bundle
catalogs to load whenever it completed the function, whether the archive
was able to be loaded or not.

Now it uses `AZStd::erase_if` to only remove archives from the list of
archives with catalog bundles when they are actually loaded.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>