Show search results in Asset Browser Thumbnail View (#14003)
Refactor a bit the underlying model to derive from QIdentityProxyModel and
using roles for special functions such as visibility of single  items when
filtering with a search string.

After the refactoring also #14002 is fixed.

Add a context menu when showing search results, to allow navigating to
the parent folder of a result item.

Signed-off-by: Alessandro Ambrosano <1006222+aambrosano@users.noreply.github.com>