Remove the AZ_TEST_CLASS_ALLOCATOR macro
This macro is designed to mimic the behavior of AZ_CLASS_ALLOCATOR, but
just redirects to malloc/free for classes that use it. In all cases where
this was used, none of the objects' base classes override `operator new`,
so these were just overriding `operator new` to explicitly specify the
default behavior.

Signed-off-by: Chris Burel <burelc@amazon.com>