Added an explicit binary directory to the mono-default configure CMake Presets.
Previously these presets did not specify a binary directory, therefore
using these presets would default to using the binary directory of the
first inherited preset.

For example windows this would result in the monolithic build being placed in
the same folder as non-monolithic builds(`build/windows`), because the
windows_unity target has a binary directory set to `build/windows`.
his makes it so that by default the monolithic binary directory is a
different directory than the non-monolithic binary directory.
This can overridden via the `-B <binary-dir>` option when cmake
configuring.

For more information see the
[configure_preset](https://cmake.org/cmake/help/latest/manual/cmake-presets.7.html?highlight=presets#configure-preset) inherits entry documentation for inheritance of presets work.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>