Up the connection check time for the EBus Connection Policy test
The `EBus.ConnectionPolicy_WaitOnSecondHandlerWhileStillLocked_CantComplete` test checked that a connection to the `BusWithConnectionPolicyUnlocksBus` EBus occured within 100 milliseconds which is a bit to low on the Jenkins Windows server machine running CI/CD.

The test failed previously due to it taking 100.133100 milliseconds for the
`connectHandler` to connect to the `BusWithConnectionPolicyUnlocksBus`
EBus.

Therefore the connection check time has been updated to 1 second.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>