Material Canvas: Fixing base and standard PBR material templates
This change fixes and cleans up all of the shaders provided with the current material canvas templates. When they were last submitted with support for transparency, everything appeared to work correctly on the tested pipelines but the depth shader was spamming warnings and asserts about requiring an SRG that was not bound. This turned out to be the pass SRG, which is not used in the depth shader code but was referenced in the included, common lighting code. The resulting log spam eventually caused a significant frame rate drop and potential crash.

Signed-off-by: gadams3 <guthadam@amazon.com>