Material Canvas: Restoring graph model caching and minor graph canvas optimizations
Material canvas does extensive traversing and sorting of nodes for populating the inspector and ordering instructions to generate code. Most accessor functions require looking up data on the graph object, which contains all of the node and connection data, which can be expensive.

This PR reimplements caching that saves parent nodes, connections, and data for some other recursive functions needed during sorting. The data is also protected using mutexes because a subsequent change makes material canvas sort and generate data asynchronously.

There are also minor changes to graph canvas to reserve space and avoid copying data in some places. During a stress test that loads a large graph of hundreds of nodes, it looks like most of the time is going to looking up and applying styles to graph canvas objects. Looking up matching styles is expensive because it copies containers of selector objects that are cloned and reallocated every time the search is done per object. This could be optimized by creating a set of functions that pass in the entity ID to be compared instead of copying the containers.

Signed-off-by: gadams3 <guthadam@amazon.com>