Switch to using a persistent resource for the projected shadow map atlas (#14247)
Cached shadows were implemented with the expectation that the shadow maps were persistent resources, however they were not. The fact that they've mostly worked thus far is surprising, but it is possible for the cached shadows to be overwritten by other passes since they're transient resources. This PR updates the resource to be persistent so that the memory won't be stomped on by some other pass.