Make it easier for users to understand static vs dynamic colliders (#14261)
**Note**
This work has been divided into multiple PRs to make it easier to review. Merging each PR into the staging branch `PhysX_StaticRigidBody_Staging_o3de`.

**Other PRs related to this change**
- https://github.com/o3de/o3de/pull/14262
- https://github.com/o3de/o3de/pull/14264
- https://github.com/o3de/o3de/pull/14267
- https://github.com/o3de/o3de/pull/14268
- https://github.com/o3de/o3de/pull/14295

Signed-off-by: moraaar <moraaar@amazon.com>

## What does this PR do?

It's not clear to users when they can move an entity or not if there is a collider in it. There is an internal physics logic that is not explicitly said, which is "a collider with no rigid body is an static collider and the user should not try to move it either by code or script canvas". A usual mistake users do is to create an entity with just a collider and move it in script canvas, which is wrong, but there is nothing in the component that suggests that's wrong.

This PR adds the new editor component "Static Rigid Body" and makes Colliders require either an static or a dynamic rigid body. This way if the user chooses to add a static rigid body it's explicitly chosen to be static, meaning it should not move. If the customer chose to add dynamic rigid body, then it's understood it can be moved.

**List of modifications**
- Added new "PhysX Static Rigid Body" component to the editor and renamed "PhysX Rigid Body" to "PhysX Dynamic Rigid Body". "PhysicsRigidBodyService" now means any rigid body service, and in addition, each one will offer specific services "PhysicsStaticRigidBodyService" and "PhysicsDynamicRigidBodyService".
- Character Controller functions internally like a Kinematic Dynamic Rigid Body, so its services have been updated to reflect it.
- PhysX Colliders now require to add either a static or dynamic rigid body component.
- RigidBody, StaticRigidBody and CharacterController components now create the simulated bodies in `OnEntityActivated`, after all the collider components have created the shapes during activation.
- The new StaticRigidBody component (same as RigidBody) implements `RigidBodyRequests` and it sends `RigidBodyNotifications` to indicate when the rigid body is enabled/disabled. These notification will be used by other components that need to know when the simulated bodies are ready (https://github.com/o3de/o3de/pull/14262)
- Fixed bug where MeshCollider component might not load the asset in time, now blocking until it's loaded since it's required during activation to create the shapes.

**Out of Scope**
- Improve MeshCollider component to not block loading its asset. This can be done in the future by making colliders notify to rigid bodies of new/modified shapes and rigid bodies to be able to listen to those notifications and recreate themselves when that happens, but that's out of scope for this work.
- Editor collider components still own the creation of simulated bodies when using EditorStaticRigidBody component. So there is still code duplication in Collider and Shape Collider classes. This could be refactored out of the colliders and moved to the new `EditorStaticRigidBody` component in the future.

## How was this PR tested?

PhysX unit tests pass (https://github.com/o3de/o3de/pull/14267).
Passed AR in local branch `PhysX_StaticRigidBody` which has all the PRs combined.
Manually tested in the editor that colliders require a rigid body and tested simple scenarios with colliders and rigid bodies.